CCCP-3646: Update the code to consider categories
The code has been updated to consider more granuality. If the variable
upgrade_categories is defined then only the categories defined in that
variable are upgraded. For example, if we plan to upgrade all, then
upgrade_categories="all" but if we plan to upgrade only bios and iDrac,
then upgrade_categories="iDRAC,BIOS"