Automate answering to bootstrap.cgi script
If keys are not yet imported into rpm and gpg, bootstrap.cgi will asks
the user if they want the script to do it. At the moment, no answer is
provided automatically, thus the playbook gets stuck.
If dell_dsu_rollout_gpg_keys is set to true, then the user wants to
do the imports through the playbook, therefore the automated reply to
the script should be negative.
If dell_dsu_rollout_gpg_keys is set to false, then we can assume that
the user is fine with letting the script to do the imports, therefore
the automated reply should be positive.